refactor: create NavAvatar and UserActionMenu components and move repeated logic into them
